Replace mutt-ezmlm-get and mutt-ezmlm-save by the much more capable mutt-ezmlm
The old scripts were getting the emails from the ezmlm email interface
(sending 'get' emails, and then extracting the attached messages in the
reply). It turned out that those emails were lacking a lot of headers,
including the very important for threading 'In-Reply-To' and 'References'.

The new script uses the NTTP interface to fetch the emails. They have
basically all the headers from the mail originally sent to the list.